feat(CB): multiple central managers
Adding plugin support for spawning separate scans without interruption.

A new client identifier parameter is now sent over the method channel.
If the client id does not exist on the platform, a new central manager is
created and stored. `PeripheralSessionIdentifier` stores both the peripheral
identifier as well as the client identifier so that any communication with
the platform related to a peripheral is associated with the same manager.